+++ security/core/src/main/java/org/seedstack/seed/security/SecurityConfig.java
@@ -16,13 +16,26 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import org.apache.shiro.authc.Authenticator;
+import org.apache.shiro.authc.pam.AuthenticationStrategy;
+import org.apache.shiro.authc.pam.FirstSuccessfulStrategy;
+import org.apache.shiro.authc.pam.ModularRealmAuthenticator;
+import org.apache.shiro.cache.CacheManager;
+import org.apache.shiro.cache.MemoryConstrainedCacheManager;
+import org.apache.shiro.mgt.SessionStorageEvaluator;
+import org.apache.shiro.mgt.SubjectDAO;
+import org.apache.shiro.mgt.SubjectFactory;
+import org.apache.shiro.subject.SubjectContext;
 import org.seedstack.coffig.Config;
 import org.seedstack.coffig.SingleValue;
+import org.seedstack.seed.security.internal.SeedSessionStorageEvaluator;
 
 @Config("security")
 public class SecurityConfig {
     private SessionConfig sessions = new SessionConfig();
     private CacheConfig cache = new CacheConfig();
+    private AuthenticationConfig authentication = new AuthenticationConfig();
+    private SubjectConfig subject = new SubjectConfig();
     private List<RealmConfig> realms = new ArrayList<>();
     private Map<String, UserConfig> users = new HashMap<>();
     private Map<String, Set<String>> roles = new HashMap<>();
@@ -34,6 +47,14 @@ public class SecurityConfig {
 
     public CacheConfig cache() {
         return cache;
+    }
+
+    public AuthenticationConfig authentication() {
+        return authentication;
+    }
+
+    public SubjectConfig subject() {
+        return subject;
     }
 
     public List<RealmConfig> getRealms() {
@@ -139,6 +160,7 @@ public class SecurityConfig {
         @SingleValue
         private boolean enabled;
         private long timeout = 1000 * 60 * 15;
+        private Class<? extends SessionStorageEvaluator> storageEvaluator = SeedSessionStorageEvaluator.class;
 
         public boolean isEnabled() {
             return enabled;
@@ -157,6 +179,76 @@ public class SecurityConfig {
             this.timeout = timeout * 1000;
             return this;
         }
+
+        public Class<? extends SessionStorageEvaluator> getStorageEvaluator() {
+            return storageEvaluator;
+        }
+
+        public SessionConfig setStorageEvaluator(Class<? extends SessionStorageEvaluator> storageEvaluator) {
+            this.storageEvaluator = storageEvaluator;
+            return this;
+        }
+    }
+
+    @Config("subject")
+    public static class SubjectConfig {
+        private Class<? extends SubjectDAO> dao;
+        private Class<? extends SubjectContext> context;
+        private Class<? extends SubjectFactory> factory;
+
+        public Class<? extends SubjectDAO> getDao() {
+            return dao;
+        }
+
+        public SubjectConfig setDao(Class<? extends SubjectDAO> dao) {
+            this.dao = dao;
+            return this;
+        }
+
+        public Class<? extends SubjectContext> getContext() {
+            return context;
+        }
+
+        public SubjectConfig setContext(Class<? extends SubjectContext> context) {
+            this.context = context;
+            return this;
+        }
+
+        public Class<? extends SubjectFactory> getFactory() {
+            return factory;
+        }
+
+        public SubjectConfig setFactory(Class<? extends SubjectFactory> factory) {
+            this.factory = factory;
+            return this;
+        }
+    }
+
+    @Config("authentication")
+    public static class AuthenticationConfig {
+        @SingleValue
+        private Class<? extends AuthenticationStrategy> strategy = FirstSuccessfulStrategy.class;
+        private Class<? extends Authenticator> authenticator = ModularRealmAuthenticator.class;
+
+        public Class<? extends Authenticator> getAuthenticator() {
+            return authenticator;
+        }
+
+        public AuthenticationConfig setAuthenticator(Class<? extends Authenticator> authenticator) {
+            this.authenticator = authenticator;
+            return this;
+        }
+
+        public Class<? extends AuthenticationStrategy> getStrategy() {
+            return strategy;
+        }
+
+        public AuthenticationConfig setStrategy(
+                Class<? extends AuthenticationStrategy> strategy) {
+            this.strategy = strategy;
+            return this;
+        }
+
     }
 
     @Config("cache")
@@ -165,6 +257,7 @@ public class SecurityConfig {
         private boolean enabled = true;
         private ItemCacheConfig authentication = new ItemCacheConfig();
         private ItemCacheConfig authorization = new ItemCacheConfig();
+        private Class<? extends CacheManager> manager = MemoryConstrainedCacheManager.class;
 
         public boolean isEnabled() {
             return enabled;
@@ -181,6 +274,15 @@ public class SecurityConfig {
 
         public ItemCacheConfig authorization() {
             return authorization;
+        }
+
+        public Class<? extends CacheManager> getManager() {
+            return manager;
+        }
+
+        public CacheConfig setManager(Class<? extends CacheManager> manager) {
+            this.manager = manager;
+            return this;
         }
 
         public static class ItemCacheConfig {