+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 * [new] A `@RequiresCrudPermissions` annotation allows to add permission checks based on the detected CRUD action of the called method.
 * [new] SPI `CrudActionResolver` has been added to security to allow for resolving the CRUD action of a particular method.
+* [new] Provides the ability to configure some Shiro implementation classes: `SubjectDAO`, `SubjectFactory`, `SubjectContext`, `Authenticator`, `AuthenticationStrategy`, `CacheManager` and `SessionStorageEvaluator`.  
 * [new] A JAX-RS implementation of `CrudActionResolver` detects the CRUD action based upon the JAX-RS annotations.
+* [fix] Make `SimplePrincipalProvider` serializable. 
 * [brk] Data import/export API has been moved to business framework.
 
 # Version 3.3.1 (2017-09-06)